Struct datafusion_row::accessor::RowAccessor

pub struct RowAccessor<'a> { /* private fields */ }

Provides read/write/modify access to a tuple stored in Row format at data[base_offset..]

Set / Update data
    in [u8]
     ─ ─ ─ ─ ─ ─ ─ ┐         Read data out as native
    │                         types or ScalarValues
                   │
    │  ┌───────────────────────┐
       │                       │
    └ ▶│         [u8]          │─ ─ ─ ─ ─ ─ ─ ─▶
       │                       │
       └───────────────────────┘

Implementations

impl<'a> RowAccessor<'a>

pub fn new(schema: &Schema, row_type: RowType) -> Self

new

pub fn new_from_layout(layout: Arc<RowLayout>) -> Self

pub fn point_to(&mut self, offset: usize, data: &'a mut [u8])

Update this row to point to position offset in base

pub fn get_bool_opt(&self, idx: usize) -> Option<bool>

pub fn get_u8_opt(&self, idx: usize) -> Option<u8>

pub fn get_u16_opt(&self, idx: usize) -> Option<u16>

pub fn get_u32_opt(&self, idx: usize) -> Option<u32>

pub fn get_u64_opt(&self, idx: usize) -> Option<u64>

pub fn get_i8_opt(&self, idx: usize) -> Option<i8>

pub fn get_i16_opt(&self, idx: usize) -> Option<i16>

pub fn get_i32_opt(&self, idx: usize) -> Option<i32>

pub fn get_i64_opt(&self, idx: usize) -> Option<i64>

pub fn get_f32_opt(&self, idx: usize) -> Option<f32>

pub fn get_f64_opt(&self, idx: usize) -> Option<f64>

pub fn get_bool_scalar(&self, idx: usize) -> ScalarValue

pub fn get_u8_scalar(&self, idx: usize) -> ScalarValue

pub fn get_u16_scalar(&self, idx: usize) -> ScalarValue

pub fn get_u32_scalar(&self, idx: usize) -> ScalarValue

pub fn get_u64_scalar(&self, idx: usize) -> ScalarValue

pub fn get_i8_scalar(&self, idx: usize) -> ScalarValue

pub fn get_i16_scalar(&self, idx: usize) -> ScalarValue

pub fn get_i32_scalar(&self, idx: usize) -> ScalarValue

pub fn get_i64_scalar(&self, idx: usize) -> ScalarValue

pub fn get_f32_scalar(&self, idx: usize) -> ScalarValue

pub fn get_f64_scalar(&self, idx: usize) -> ScalarValue

pub fn get_as_scalar(&self, dt: &DataType, index: usize) -> ScalarValue

pub fn add_u8(&mut self, idx: usize, value: u8)

add field at idx with value

pub fn add_u16(&mut self, idx: usize, value: u16)

add field at idx with value

pub fn add_u32(&mut self, idx: usize, value: u32)

add field at idx with value

pub fn add_u64(&mut self, idx: usize, value: u64)

add field at idx with value

pub fn add_i8(&mut self, idx: usize, value: i8)

add field at idx with value

pub fn add_i16(&mut self, idx: usize, value: i16)

add field at idx with value

pub fn add_i32(&mut self, idx: usize, value: i32)

add field at idx with value

pub fn add_i64(&mut self, idx: usize, value: i64)

add field at idx with value

pub fn add_f32(&mut self, idx: usize, value: f32)

add field at idx with value

pub fn add_f64(&mut self, idx: usize, value: f64)

add field at idx with value

pub fn max_u8(&mut self, idx: usize, value: u8)

check max then update

pub fn max_u16(&mut self, idx: usize, value: u16)

check max then update

pub fn max_u32(&mut self, idx: usize, value: u32)

check max then update

pub fn max_u64(&mut self, idx: usize, value: u64)

check max then update

pub fn max_i8(&mut self, idx: usize, value: i8)

check max then update

pub fn max_i16(&mut self, idx: usize, value: i16)

check max then update

pub fn max_i32(&mut self, idx: usize, value: i32)

check max then update

pub fn max_i64(&mut self, idx: usize, value: i64)

check max then update

pub fn max_f32(&mut self, idx: usize, value: f32)

check max then update

pub fn max_f64(&mut self, idx: usize, value: f64)

check max then update

pub fn min_u8(&mut self, idx: usize, value: u8)

check max then update

pub fn min_u16(&mut self, idx: usize, value: u16)

check max then update

pub fn min_u32(&mut self, idx: usize, value: u32)

check max then update

pub fn min_u64(&mut self, idx: usize, value: u64)

check max then update

pub fn min_i8(&mut self, idx: usize, value: i8)

check max then update

pub fn min_i16(&mut self, idx: usize, value: i16)

check max then update

pub fn min_i32(&mut self, idx: usize, value: i32)

check max then update

pub fn min_i64(&mut self, idx: usize, value: i64)

check max then update

pub fn min_f32(&mut self, idx: usize, value: f32)

check max then update

pub fn min_f64(&mut self, idx: usize, value: f64)

check max then update